Added error handling for Sync and Browser Refresh utils
